+++ src/main/kotlin/com/itsci/mju/maebanjumpen/partyrole/dto/AccountManagerDTO.kt
@@ -1,7 +1,7 @@
 package com.itsci.mju.maebanjumpen.partyrole.dto
 
 class AccountManagerDTO : PartyRoleDTO() {
-    var managerID: Int? = null
+    var managerID: Long? = null
 
     override fun getType(): String = "accountManager"
 }
