+++ src/main/kotlin/com/itsci/mju/maebanjumpen/partyrole/repository/AccountManagerRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AccountManagerRepository : JpaRepository<AccountManager, Int>
+interface AccountManagerRepository : JpaRepository<AccountManager, Long>
