+++ src/main/kotlin/com/itsci/mju/maebanjumpen/partyrole/service/AccountManagerService.kt
@@ -4,9 +4,9 @@ import com.itsci.mju.maebanjumpen.partyrole.dto.AccountManagerDTO
 
 interface AccountManagerService {
     fun listAllAccountManagers(): List<AccountManagerDTO>
-    fun getAccountManagerById(id: Int): AccountManagerDTO
+    fun getAccountManagerById(id: Long): AccountManagerDTO
     fun createAccountManager(accountManager: AccountManagerDTO): AccountManagerDTO
-    fun deleteAccountManager(id: Int)
-    fun updateAccountManager(id: Int, accountManager: AccountManagerDTO): AccountManagerDTO
+    fun deleteAccountManager(id: Long)
+    fun updateAccountManager(id: Long, accountManager: AccountManagerDTO): AccountManagerDTO
 }
