+++ src/main/kotlin/com/itsci/mju/maebanjumpen/partyrole/service/HirerService.kt
@@ -4,11 +4,11 @@ import com.itsci.mju.maebanjumpen.partyrole.dto.HirerDTO
 
 interface HirerService {
     fun saveHirer(hirerDto: HirerDTO): HirerDTO
-    fun getHirerById(id: Int): HirerDTO
+    fun getHirerById(id: Long): HirerDTO
     fun getAllHirers(): List<HirerDTO>
-    fun updateHirer(id: Int, hirerDto: HirerDTO): HirerDTO
-    fun deleteHirer(id: Int)
-    fun deductBalance(hirerId: Int, amount: Double)
-    fun addBalance(hirerId: Int, amount: Double)
+    fun updateHirer(id: Long, hirerDto: HirerDTO): HirerDTO
+    fun deleteHirer(id: Long)
+    fun deductBalance(hirerId: Long, amount: Double)
+    fun addBalance(hirerId: Long, amount: Double)
 }
