+++ src/main/kotlin/com/itsci/mju/maebanjumpen/partyrole/service/HousekeeperService.kt
@@ -5,13 +5,13 @@ import com.itsci.mju.maebanjumpen.partyrole.dto.HousekeeperDTO
 
 interface HousekeeperService {
     fun getAllHousekeepers(): List<HousekeeperDTO>
-    fun getHousekeeperDetailById(id: Int): HousekeeperDetailDTO?
+    fun getHousekeeperDetailById(id: Long): HousekeeperDetailDTO?
     fun saveHousekeeper(housekeeperDto: HousekeeperDTO): HousekeeperDTO
-    fun updateHousekeeper(id: Int, housekeeperDto: HousekeeperDTO): HousekeeperDTO
-    fun deleteHousekeeper(id: Int)
-    fun calculateAndSetAverageRating(housekeeperId: Int)
-    fun addBalance(housekeeperId: Int, amount: Double)
-    fun deductBalance(housekeeperId: Int, amount: Double)
+    fun updateHousekeeper(id: Long, housekeeperDto: HousekeeperDTO): HousekeeperDTO
+    fun deleteHousekeeper(id: Long)
+    fun calculateAndSetAverageRating(housekeeperId: Long)
+    fun addBalance(housekeeperId: Long, amount: Double)
+    fun deductBalance(housekeeperId: Long, amount: Double)
     fun getHousekeepersByStatus(status: String): List<HousekeeperDTO>
     fun getNotVerifiedOrNullStatusHousekeepers(): List<HousekeeperDTO>
 }
