+++ src/main/kotlin/com/itsci/mju/maebanjumpen/partyrole/service/MemberService.kt
@@ -5,10 +5,10 @@ import java.util.Optional
 
 interface MemberService {
     fun saveMember(memberDto: MemberDTO): MemberDTO
-    fun getMemberById(id: Int): Optional<MemberDTO>
+    fun getMemberById(id: Long): Optional<MemberDTO>
     fun getAllMembers(): List<MemberDTO>
-    fun updateMember(id: Int, memberDto: MemberDTO): MemberDTO
-    fun deleteMember(id: Int)
-    fun deductBalance(memberId: Int, amount: Double): MemberDTO
+    fun updateMember(id: Long, memberDto: MemberDTO): MemberDTO
+    fun deleteMember(id: Long)
+    fun deductBalance(memberId: Long, amount: Double): MemberDTO
 }
