+++ src/main/kotlin/com/itsci/mju/maebanjumpen/penalty/repository/PenaltyRepository.kt
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PenaltyRepository : JpaRepository<Penalty, Int>
+interface PenaltyRepository : JpaRepository<Penalty, Long>
