+++ src/main/kotlin/com/itsci/mju/maebanjumpen/transaction/dto/QrCodeRequestDTO.kt
@@ -1,7 +1,7 @@
 package com.itsci.mju.maebanjumpen.transaction.dto
 
 data class QrCodeRequestDTO(
-    var memberId: Int? = null,
+    var memberId: Long? = null,
     var amount: Double? = null
 )
