+++ src/main/kotlin/com/itsci/mju/maebanjumpen/transaction/service/TransactionService.kt
@@ -6,12 +6,12 @@ import java.util.Optional
 
 interface TransactionService {
     fun getAllTransactions(): List<TransactionDTO>
-    fun getTransactionById(id: Int): Optional<TransactionDTO>
+    fun getTransactionById(id: Long): Optional<TransactionDTO>
     fun saveTransaction(transactionDto: TransactionDTO): TransactionDTO
-    fun deleteTransaction(id: Int)
-    fun getTransactionsByMemberId(memberId: Int): List<TransactionDTO>
+    fun deleteTransaction(id: Long)
+    fun getTransactionsByMemberId(memberId: Long): List<TransactionDTO>
     fun getWithdrawalRequests(): List<TransactionDTO>
-    fun updateWithdrawalRequestStatus(transactionId: Int, newStatus: String): Optional<TransactionDTO>
+    fun updateWithdrawalRequestStatus(transactionId: Long, newStatus: String): Optional<TransactionDTO>
     fun processOmiseChargeComplete(omisePayload: JsonNode)
     fun getLocalizedStatus(status: String?, isEnglish: Boolean): String
     fun getStatusColorHex(status: String?): String
